contracts inits

+++ contracts/src/interfaces/IERC20.sol
@@ -2,8 +2,8 @@
 pragma solidity ^0.8.20;
 
 interface IERC20 {
-    function transferFrom(address from, address to, uint value) external returns (bool);
-    function transfer(address to, uint value) external returns (bool);
-    function approve(address spender, uint value) external returns (bool);
-    function balanceOf(address account) external view returns (uint);
+    function transferFrom(address from, address to, uint256 value) external returns (bool);
+    function transfer(address to, uint256 value) external returns (bool);
+    function approve(address spender, uint256 value) external returns (bool);
+    function balanceOf(address account) external view returns (uint256);
 }
